Got permissions working

+++ app/src/server/router/homes.ts
@@ -2,6 +2,8 @@ import { createProtectedRouter } from "./context";
 import { z } from "zod";
 import { getHomesByUserId, canUserViewHome } from "../db/HomeService";
 import { getSignedImage } from "./image-upload";
+import { hasPermission } from "server/db/UserService";
+import { Permission } from "types/permissions";
 
 export const homesRouter = createProtectedRouter()
     .query("getHomes", {
@@ -67,13 +69,29 @@ export const homesRouter = createProtectedRouter()
             //Creates image key here so the image can bey placed in the S3 Bucket and referenced in the Prisma Database
 
             //Puts home info in prisma database
-            return await ctx.prisma.home.create({
+            const home =  await ctx.prisma.home.create({
                 data: {
                     name: input.name,
                     image: input.image,
                     address: input.address,
                 },
             });
+            await ctx.prisma.occupies.create({
+                data: {
+                    userId: ctx.session.user.id,
+                    homeId: home.id,
+                }
+            });
+
+            await ctx.prisma.permission.create({
+                data: {
+                    occupiesUserId: ctx.session.user.id,
+                    occupiesHomeId: home.id,
+                    name: Permission.Owner,
+                }
+            })
+
+            return home;
         },
     })
     .mutation("updateHome", {
@@ -84,6 +102,9 @@ export const homesRouter = createProtectedRouter()
             address: z.string(),
         }),
         async resolve({ ctx, input }) {
+            if(!(await hasPermission(ctx.session.user.id, input.id, Permission.Edit, ctx.prisma)) ||
+               !(await canUserViewHome(ctx.session.user.id,input.id, ctx.prisma)))
+                return;
             return await ctx.prisma.home.update({
                 where: {
                     id: input.id,
@@ -101,6 +122,9 @@ export const homesRouter = createProtectedRouter()
             id: z.string(),
         }),
         async resolve({ ctx, input }) {
+            if(!(await hasPermission(ctx.session.user.id, input.id, Permission.Delete, ctx.prisma)) ||
+               !(await canUserViewHome(ctx.session.user.id,input.id, ctx.prisma)))
+                return;
             await ctx.prisma.occupies.deleteMany({
                 where: {
                     homeId: input.id,
